+++ defLob.go
@@ -12,12 +12,10 @@ package ora
 import "C"
 import (
 	"bytes"
-	"encoding/binary"
 	"fmt"
 	"io"
 	"sync"
 	"sync/atomic"
-	"unicode/utf16"
 	"unsafe"
 )
 
@@ -35,7 +33,6 @@ type defLob struct {
 	gct  GoColumnType
 	sqlt C.ub2
 	lobs []*C.OCILobLocator
-	sync.Mutex
 }
 
 func (def *defLob) define(position int, sqlt C.ub2, gct GoColumnType, rset *Rset) error {
@@ -52,9 +49,11 @@ func (def *defLob) define(position int, sqlt C.ub2, gct GoColumnType, rset *Rset
 	env := rset.stmt.ses.srv.env
 	//rset.RUnlock()
 	def.lobs = (*((*[MaxFetchLen]*C.OCILobLocator)(C.malloc(C.size_t(fetchLen) * C.sof_LobLocatorp))))[:fetchLen]
+	def.ensureAllocatedLength(len(def.lobs))
 	if err := def.ociDef.defineByPos(position, unsafe.Pointer(&def.lobs[0]), int(C.sof_LobLocatorp), int(sqlt)); err != nil {
 		return err
 	}
+	// https://docs.oracle.com/cd/B28359_01/appdev.111/b28395/oci07lob.htm#CHDDHFAB
 	prefetchLength := C.boolean(C.TRUE)
 	return env.setAttr(unsafe.Pointer(def.ocidef), C.OCI_HTYPE_DEFINE,
 		unsafe.Pointer(&prefetchLength), 0, C.OCI_ATTR_LOBPREFETCH_LENGTH)
@@ -114,6 +113,7 @@ func (def *defLob) Reader(offset int) io.ReadCloser {
 	}
 	//def.rset.RUnlock()
 	def.lobs[offset] = nil // don't use it anywhere else
+	def.allocated[offset] = false
 	def.Unlock()
 	//fmt.Printf("%p.Reader(%d): %p\n", def, offset, lr)
 	return lr
@@ -176,6 +176,7 @@ func (def *defLob) alloc() error {
 	//def.rset.RUnlock()
 	ocienv := env.ocienv
 	for i := range def.lobs {
+		def.allocated[i] = false
 		r := C.OCIDescriptorAlloc(
 			unsafe.Pointer(ocienv),                          //CONST dvoid   *parenth,
 			(*unsafe.Pointer)(unsafe.Pointer(&def.lobs[i])), //dvoid         **descpp,
@@ -187,6 +188,7 @@ func (def *defLob) alloc() error {
 		} else if r == C.OCI_INVALID_HANDLE {
 			return errNew("unable to allocate oci lob handle during define")
 		}
+		def.allocated[i] = true
 	}
 	return nil
 }
@@ -194,15 +196,17 @@ func (def *defLob) alloc() error {
 func (def *defLob) free() {
 	def.Lock()
 	defer def.Unlock()
-	def.arrHlp.close()
 	ses := def.rset.stmt.ses
 	for i, lob := range def.lobs {
 		if lob == nil {
 			continue
 		}
 		def.lobs[i] = nil
-		lobClose(ses, lob)
+		if def.allocated[i] {
+			lobClose(ses, lob)
+		}
 	}
+	def.arrHlp.close()
 }
 
 func (def *defLob) close() (err error) {
@@ -229,7 +233,6 @@ type lobReader struct {
 	ses           *Ses
 	ociLobLocator *C.OCILobLocator
 	piece, csfrm  C.ub1
-	csid          C.ub2
 	off           C.oraub8
 	opened        bool
 
@@ -279,9 +282,7 @@ func (lr *lobReader) Read(p []byte) (n int, err error) {
 		// Open the lob to obtain length; round-trip to database
 		//Log.Infof("Reader OCILobOpen %p", def.ociLobLocator)
 		//fmt.Printf("lobOpen(%p loc=%p)\n", lr, lr.ociLobLocator)
-		lr.Length, lr.csid, lr.csfrm, err = lobOpen(ses, ociLobLocator, C.OCI_LOB_READONLY)
-		lr.csid = C.ub2(atomic.LoadUint32(&csIDAl32UTF8))
-
+        lr.Length, err = lobOpen(ses, ociLobLocator, C.OCI_LOB_READONLY)
 		lr.Unlock()
 		if err != nil {
 			return 0, err
@@ -290,7 +291,7 @@ func (lr *lobReader) Read(p []byte) (n int, err error) {
 
 	lr.Lock()
 	Length, off := lr.Length, lr.off
-	piece, csid, csfrm := lr.piece, lr.csid, lr.csfrm
+	piece := lr.piece
 	lr.Unlock()
 	if Length == 0 || off >= Length {
 		return 0, io.EOF
@@ -310,8 +311,8 @@ func (lr *lobReader) Read(p []byte) (n int, err error) {
 		piece,                 //ub1                piece,
 		nil,                   //void               *ctxp,
 		nil,                   //OCICallbackLobRead2 (cbfp)
-		csid,                  //C.ub2(0),              //ub2                csid,
-		csfrm,                 //lr.charsetForm,                          //ub1                csfrm );
+		C.ub2(atomic.LoadUint32(&csIDAl32UTF8)), //ub2             csid,
+		C.SQLCS_IMPLICIT,                        //lr.charsetForm,                          //ub1                csfrm );
 	)
 	//Log.Infof("LobRead2 returned %d amt=%d", r, byteAmt)
 	err = nil
@@ -324,21 +325,9 @@ func (lr *lobReader) Read(p []byte) (n int, err error) {
 	case C.OCI_INVALID_HANDLE:
 		err = fmt.Errorf("Invalid handle %v", ociLobLocator)
 	}
-	ses.logF(_drv.Cfg().Log.Ses.Close, "OCILobRead2(%p) off=%d amt=%d csid=%d csfrm=%d err=%v\n", ociLobLocator, off, byteAmt, csid, csfrm, err)
+	ses.logF(_drv.Cfg().Log.Ses.Close, "OCILobRead2(%p) off=%d amt=%d err=%v\n", ociLobLocator, off, byteAmt, err)
+	lr.off += byteAmt
 	// byteAmt represents the amount copied into buffer by oci
-	if byteAmt != 0 {
-		lr.Lock()
-		lr.off += byteAmt
-		lr.piece = C.OCI_NEXT_PIECE
-		lr.Unlock()
-
-		if csid == 2000 && byteAmt > 1 {
-			// UTF-16
-			u16 := make([]uint16, int(byteAmt/2))
-			binary.Read(bytes.NewReader(p[:int(byteAmt)]), binary.BigEndian, &u16)
-			return copy(p[:], string(utf16.Decode(u16))), err
-		}
-	}
 	return int(byteAmt), err
 }
 
@@ -376,8 +365,6 @@ type lobReadWriter struct {
 	ses           *Ses
 	ociLobLocator *C.OCILobLocator
 	size          C.oraub8
-	csid          C.ub2
-	csfrm         C.ub1
 }
 
 // Size returns the actual size of the LOB.
@@ -410,9 +397,6 @@ func (lrw *lobReadWriter) Truncate(length int64) error {
 
 // ReadAt reads into p, starting from off.
 func (lrw *lobReadWriter) ReadAt(p []byte, off int64) (n int, err error) {
-	if lrw.csid == 0 {
-		lrw.csid = C.ub2(atomic.LoadUint32(&csIDAl32UTF8))
-	}
 	byteAmt := C.oraub8(len(p))
 	//Log.Infof("LobRead2 off=%d amt=%d", off, len(p))
 	r := C.OCILobRead2(
@@ -427,8 +411,8 @@ func (lrw *lobReadWriter) ReadAt(p []byte, off int64) (n int, err error) {
 		C.OCI_ONE_PIECE,        //ub1                piece,
 		nil,                    //void               *ctxp,
 		nil,                    //OCICallbackLobRead2 (cbfp)
-		lrw.csid,               //ub2                csid,
-		lrw.csfrm,              //ub1                csfrm );
+		C.ub2(atomic.LoadUint32(&csIDAl32UTF8)),
+		C.SQLCS_IMPLICIT, //ub1                csfrm );
 	)
 	//Log.Infof("LobRead2 returned %d amt=%d", r, byteAmt)
 	switch r {
@@ -474,7 +458,7 @@ func (lrw *lobReadWriter) WriteAt(p []byte, off int64) (n int, err error) {
 }
 
 func lobOpen(ses *Ses, lob *C.OCILobLocator, mode C.ub1) (
-	length C.oraub8, csid C.ub2, csfrm C.ub1, err error,
+	length C.oraub8, err error,
 ) {
 	ses.RLock()
 	ocisvcctx := ses.ocisvcctx
@@ -506,7 +490,7 @@ func lobOpen(ses *Ses, lob *C.OCILobLocator, mode C.ub1) (
 	//Log.Infof("OCILobOpen %p returned %d", lob, r)
 	if r != C.OCI_SUCCESS {
 		lobClose(ses, lob)
-		return 0, csid, csfrm, env.ociError("OCILobOpen")
+		return 0, env.ociError("OCILobOpen")
 	}
 	// get the length of the lob
 	// For character LOBs, it is the number of characters; for binary LOBs and BFILEs,
@@ -518,34 +502,13 @@ func lobOpen(ses *Ses, lob *C.OCILobLocator, mode C.ub1) (
 		&length,    //oraub8 *lenp)
 	); r == C.OCI_ERROR {
 		lobClose(ses, lob)
-		return length, csid, csfrm, env.ociError("OCILobGetLength2")
+		return length, env.ociError("OCILobGetLength2")
 	}
 
-	if r = C.OCILobCharSetId(
-		env.ocienv,
-		env.ocierr,
-		lob,
-		&csid,
-	); r == C.OCI_ERROR {
-		lobClose(ses, lob)
-		return length, csid, csfrm, env.ociError("OCILobCharSetId")
-	}
-	if csid == 0 {
-		csid = C.ub2(atomic.LoadUint32(&csIDAl32UTF8))
-	}
-	if r = C.OCILobCharSetForm(
-		env.ocienv,
-		env.ocierr,
-		lob,
-		&csfrm,
-	); r == C.OCI_ERROR {
-		lobClose(ses, lob)
-		return length, csid, csfrm, env.ociError("OCILobCharSetForm")
-	}
-	return length, csid, csfrm, nil
+	return length, nil
 }
 
-func lobClose(ses *Ses, lob *C.OCILobLocator) error {
+func lobClose(ses *Ses, lob *C.OCILobLocator) (err error) {
 	if lob == nil {
 		return nil
 	}