+++ defUint64.go
@@ -57,7 +57,7 @@ func (def *defUint64) value(offset int) (value interface{}, err error) {
 }
 
 func (def *defUint64) alloc() error { return nil }
-func (def *defUint64) free()        {
+func (def *defUint64) free() {
 	def.arrHlp.close()
 }