+++ stmtCfg.go
@@ -12,13 +12,14 @@ package ora
 // StmtCfg is immutable, so every Set method returns a new
 // instance, maybe with Err set, too.
 type StmtCfg struct {
-	prefetchRowCount    uint32
-	prefetchMemorySize  uint32
-	longBufferSize      uint32
-	longRawBufferSize   uint32
-	lobBufferSize       int
-	stringPtrBufferSize int
-	byteSlice           GoColumnType
+	prefetchRowCount      uint32
+	prefetchMemorySize    uint32
+	longBufferSize        uint32
+	longRawBufferSize     uint32
+	lobBufferSize         int
+	stringPtrBufferSize   int
+	fetchLen, lobFetchLen int
+	byteSlice             GoColumnType
 
 	// IsAutoCommitting determines whether DML statements are automatically
 	// committed.
@@ -56,6 +57,8 @@ type StmtCfg struct {
 // NewStmtCfg returns a StmtCfg with default values.
 func NewStmtCfg() StmtCfg {
 	var c StmtCfg
+	c.fetchLen = DefaultFetchLen
+	c.lobFetchLen = DefaultLOBFetchLen
 	c.prefetchRowCount = 128
 	c.prefetchMemorySize = 128 << 20 // 134,217,728
 	c.longBufferSize = 16 << 20      // 16,777,216
@@ -256,6 +259,46 @@ func (c StmtCfg) SetByteSlice(gct GoColumnType) StmtCfg {
 // if the destination column is NUMBER, BINARY_DOUBLE, BINARY_FLOAT or FLOAT.
 func (c StmtCfg) ByteSlice() GoColumnType {
 	return c.byteSlice
+}
+
+// returns a value of the lobFetchLen
+func (c StmtCfg) LOBFetchLen() int {
+	return c.lobFetchLen
+}
+
+// returns a value of the fetchLen
+func (c StmtCfg) FetchLen() int {
+	return c.fetchLen
+}
+
+// SetFetchLen overrides DefaultFetchLen for prefetch lengths.
+func (c StmtCfg) SetFetchLen(length int) StmtCfg {
+	if length <= 0 {
+		return c
+	}
+	if length >= MaxFetchLen {
+		length = MaxFetchLen
+	}
+	c.fetchLen = length
+	return c
+}
+
+// SetLOBFetchLen overrides DefaultLOBFetchLen for prefetch LOB lengths.
+//
+// This affects result sets with any of the following column types:
+// C.SQLT_LNG, C.SQLT_BFILE, C.SQLT_BLOB, C.SQLT_CLOB, C.SQLT_LBI
+//
+// Caution: the default buffer size for blob is 1MB. So, for example a single
+// fetch from the result set that contains just one blob will consume 128MB of RAM
+func (c StmtCfg) SetLOBFetchLen(length int) StmtCfg {
+	if length <= 0 {
+		return c
+	}
+	if length >= MaxFetchLen {
+		length = MaxFetchLen
+	}
+	c.lobFetchLen = length
+	return c
 }
 
 func (c StmtCfg) SetNumberInt(gct GoColumnType) StmtCfg {