+++ z_lob_test.go
@@ -6,6 +6,8 @@ package ora_test
 
 import (
 	"bytes"
+	"context"
+	"database/sql"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
@@ -15,9 +17,69 @@ import (
 	"strings"
 	"sync"
 	"testing"
+	"time"
 
 	ora "github.com/tianlin/ora"
 )
+
+func TestLOBCloseStatement(t *testing.T) {
+	qry := `DECLARE
+  cur1 SYS_REFCURSOR;
+  cur2 SYS_REFCURSOR;
+  cur3 SYS_REFCURSOR;
+  cur4 SYS_REFCURSOR;
+BEGIN
+  OPEN cur1 FOR
+    SELECT '1', SYSDATE dt FROM DUAL
+	UNION ALL
+	SELECT object_name, SYSDATE FROM all_objects;
+  :1 := cur1;
+  OPEN cur2 FOR
+    SELECT '1', SYSDATE dt FROM DUAL
+	UNION ALL
+	SELECT object_name, SYSDATE FROM all_objects;
+  :2 := cur2;
+  OPEN cur3 FOR
+    SELECT '1', SYSDATE dt FROM DUAL
+	UNION ALL
+	SELECT object_name, SYSDATE FROM all_objects;
+  :3 := cur3;
+  OPEN cur4 FOR
+    SELECT '2', SYSDATE dt FROM DUAL
+	UNION ALL
+	SELECT object_name, SYSDATE FROM all_objects;
+  :4 := cur4;
+END;`
+
+	testSes := getSes(t)
+	defer testSes.Close()
+
+	for _, doNext := range []bool{true, false} {
+		stmt, err := testSes.Prep(qry)
+		if err != nil {
+			t.Error(qry, err)
+		}
+
+		cur := make([]*ora.Rset, 4)
+		for i := range cur {
+			cur[i] = &ora.Rset{}
+		}
+		if _, err = stmt.Exe(cur[0], cur[1], cur[2], cur[3]); err != nil {
+			t.Error(err)
+		}
+		if !doNext {
+			stmt.Close()
+			continue
+		}
+		for _, c := range cur {
+			row := c.NextRow()
+			if row != nil {
+				fmt.Println(row)
+			}
+		}
+		stmt.Close()
+	}
+}
 
 func TestLobSelect(t *testing.T) {
 	tbl := tableName()
@@ -194,6 +256,9 @@ END test_get_json;`,
 	); err != nil {
 		t.Skipf("create function: %v", err)
 	}
+	testSes := getSes(t)
+	defer testSes.Close()
+
 	//enableLogging(t)
 	stmt, err := testSes.Prep("CALL test_get_json(:1, :2)", ora.OraBin, ora.S)
 	if err != nil {
@@ -270,6 +335,9 @@ Læringsutbyttet defineres i forhold til områder for kunnskap, ferdigheter og h
 	}
 
 	qry = "SELECT * FROM " + tbl + " WHERE emnekode = :1"
+
+	testSes := getSes(t)
+	defer testSes.Close()
 
 	enableLogging(t)
 	// LOB
@@ -585,6 +653,9 @@ Læringsutbyttet defineres i forhold til områder for kunnskap, ferdigheter og h
 		}
 	}
 
+	testSes := getSes(t)
+	defer testSes.Close()
+
 	qry = "SELECT * FROM " + tbl + " WHERE emnekode = :1"
 
 	// string
@@ -723,6 +794,9 @@ Læringsutbyttet defineres i forhold til områder for kunnskap, ferdigheter og h
 
 func TestLobIssue191(t *testing.T) {
 	enableLogging(t)
+	testSes := getSes(t)
+	defer testSes.Close()
+
 	l := &ora.Lob{}
 	stmt, err := testSes.Prep(`begin :1 := null; end;`)
 	if err != nil {
@@ -785,4 +859,90 @@ func reverseString(s string) string {
 		j--
 	}
 	return string(runes)
+}
+
+func TestLobIssue237(t *testing.T) {
+	t.Parallel()
+	//defer tl.enableLogging(t)()
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	ms, err := newMetricSet(ctx, testDb)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer ms.Close()
+
+	for i := 0; i < 100; i++ {
+		if err := ctx.Err(); err != nil {
+			break
+		}
+		events, err := ms.Fetch(ctx)
+		t.Log("events:", len(events))
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+}
+
+func newMetricSet(ctx context.Context, db *sql.DB) (*metricSet, error) {
+	qry := "select /* metricset: sqlstats */ inst_id, sql_fulltext, last_active_time from gv$sqlstats WHERE ROWNUM < 11"
+	stmt, err := db.PrepareContext(ctx, qry)
+	if err != nil {
+		return nil, err
+	}
+
+	return &metricSet{
+		stmt: stmt,
+	}, nil
+}
+
+type metricSet struct {
+	stmt *sql.Stmt
+}
+
+func (m *metricSet) Close() error {
+	st := m.stmt
+	m.stmt = nil
+	if st == nil {
+		return nil
+	}
+	return st.Close()
+}
+
+// Fetch methods implements the data gathering and data conversion to the right format
+// It returns the event which is then forward to the output. In case of an error, a
+// descriptive error must be returned.
+func (m *metricSet) Fetch(ctx context.Context) ([]event, error) {
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	rows, err := m.stmt.QueryContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var events []event
+	var buf bytes.Buffer
+	for rows.Next() {
+		var e event
+		var lob *ora.Lob
+		if err := rows.Scan(&e.ID, &lob, &e.LastActive); err != nil {
+			return events, err
+		}
+		buf.Reset()
+		if _, err := io.Copy(&buf, lob); err != nil {
+			return events, err
+		}
+		e.Text = buf.String()
+		events = append(events, e)
+	}
+
+	return events, nil
+}
+
+type event struct {
+	ID         int64
+	Text       string
+	LastActive time.Time
 }